api/requirements.txt
fastapi
uvicorn
pydantic
xgboost
bcrypt
python-jose
scikit-learn
pandas
joblib
shap
